# lib/api-zod/src/generated/types/createLawFirmRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Checker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLawFirmRequestFirmType } from "./createLawFirmRequestFirmType";

export interface CreateLawFirmRequest {
  name: string;
  firmType: CreateLawFirmRequestFirmType;
  jurisdictions?: string[];
  practiceAreas?: string[];
  contactName?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
  relationshipPartner?: string | null;
  notes?: string | null;
}