# lib/api-zod/src/generated/types/createPanelBaselineDocumentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Checker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePanelBaselineDocumentRequestDocumentKind } from "./createPanelBaselineDocumentRequestDocumentKind";
import type { CreatePanelRateItem } from "./createPanelRateItem";

export interface CreatePanelBaselineDocumentRequest {
  documentKind: CreatePanelBaselineDocumentRequestDocumentKind;
  versionLabel: string;
  fileName?: string | null;
  rates?: CreatePanelRateItem[];
}