# lib/api-zod/src/generated/types/updateLawFirmRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Checker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLawFirmRequestFirmType } from "./updateLawFirmRequestFirmType";

export interface UpdateLawFirmRequest {
  name?: string;
  firmType?: UpdateLawFirmRequestFirmType;
  jurisdictions?: string[];
  practiceAreas?: string[];
  contactName?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
  relationshipPartner?: string | null;
  notes?: string | null;
  isActive?: boolean;
}